requirements.txt
aiogram==3.15.0
asyncpg==0.30.0
python-dotenv==1.1.0
deep-translator==1.11.4
